+++ apps/api/flowaccount-sdk/src/page365/services/webhookServices.spec.ts
@@ -1,0 +1,15 @@
+import { WebhookService } from './webhookService'
+import "reflect-metadata"
+
+describe('webhookServices', () => {
+    it('Create cash invoice', async() => {
+        
+        let ws = await WebhookService
+
+        let payload = {
+
+        }
+        // expect(ws.prototype.inboundPayload(payload)).toBe(true);
+        expect(ws.prototype.testFunc('test')).toBe('ok');
+    }); 
+})
