+++ jest.config.js
@@ -1,5 +1,5 @@
-const { getJestProjects } = require('@nrwl/jest');
-
+/** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
 module.exports = {
-  projects: getJestProjects(),
+    preset: 'ts-jest',
+    testEnvironment: 'node',
 };
