+++ package.json
@@ -1,48 +1,46 @@
 {
-  "name": "flowaccount-open-api",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "start": "nx serve",
-    "build": "nx build",
-    "test": "nx test"
-  },
-  "private": true,
-  "dependencies": {
-    "@flowaccount/openapi-sdk": "^0.0.39",
-    "@types/aws-lambda": "^8.10.86",
-    "ajv": "^6.12.6",
-    "graphql": "^16.3.0",
-    "graphql-language-service-parser": "^1.10.4",
-    "graphql-request": "^4.2.0",
-    "moment": "^2.29.1",
-    "reflect-metadata": "^0.1.13",
-    "tslib": "^2.0.0",
-    "tsyringe": "^4.6.0"
-  },
-  "devDependencies": {
-    "@flowaccount/nx-serverless": "^1.1.2",
-    "@nrwl/cli": "^13.9.5",
-    "@nrwl/eslint-plugin-nx": "^13.9.5",
-    "@nrwl/jest": "^13.9.5",
-    "@nrwl/linter": "^13.9.5",
-    "@nrwl/node": "^13.9.5",
-    "@nrwl/tao": "^13.9.5",
-    "@nrwl/workspace": "^13.9.5",
-    "@openapitools/openapi-generator-cli": "^2.4.26",
-    "@trumbitta/nx-plugin-openapi": "^1.10.0",
-    "@types/jest": "^27.4.1",
-    "@types/mocha": "^9.1.0",
-    "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.10.0",
-    "@typescript-eslint/parser": "~5.10.0",
-    "eslint": "~8.7.0",
-    "eslint-config-prettier": "8.1.0",
-    "jest": "^27.5.1",
-    "mocha": "^9.2.2",
-    "prettier": "^2.5.1",
-    "serverless-offline": "^8.3.1",
-    "ts-jest": "^27.1.3",
-    "typescript": "~4.5.2"
-  }
+    "name": "flowaccount-open-api",
+    "version": "0.0.0",
+    "license": "MIT",
+    "scripts": {
+        "start": "nx serve",
+        "build": "nx build",
+        "test": "jest"
+    },
+    "private": true,
+    "dependencies": {
+        "@flowaccount/openapi-sdk": "^0.0.39",
+        "@types/aws-lambda": "^8.10.86",
+        "graphql": "^16.3.0",
+        "graphql-language-service-parser": "^1.10.4",
+        "graphql-request": "^4.2.0",
+        "moment": "^2.29.1",
+        "node-gyp": "^9.0.0",
+        "reflect-metadata": "^0.1.13",
+        "tslib": "^2.0.0",
+        "tsyringe": "^4.7.0"
+    },
+    "devDependencies": {
+        "@flowaccount/nx-serverless": "^0.5.4",
+        "@nrwl/cli": "^13.9.5",
+        "@nrwl/eslint-plugin-nx": "^13.9.5",
+        "@nrwl/linter": "^13.9.5",
+        "@nrwl/node": "^13.9.5",
+        "@nrwl/tao": "^13.9.5",
+        "@nrwl/workspace": "^13.9.5",
+        "@openapitools/openapi-generator-cli": "^2.4.26",
+        "@trumbitta/nx-plugin-openapi": "^1.10.0",
+        "@types/jest": "^27.5.1",
+        "@types/mocha": "^9.1.0",
+        "@types/node": "16.11.7",
+        "@typescript-eslint/eslint-plugin": "~5.10.0",
+        "@typescript-eslint/parser": "~5.10.0",
+        "eslint": "~8.7.0",
+        "eslint-config-prettier": "8.1.0",
+        "mocha": "^9.2.2",
+        "prettier": "^2.5.1",
+        "serverless-offline": "^8.3.1",
+        "ts-jest": "^27.1.3",
+        "typescript": "~4.5.2"
+    }
 }
